Extract job schemas into a shared package entry
