+++ apps/app/src/app/api/jobs/logs/schemas.ts
@@ -1,4 +1,4 @@
-import { listJobLogsInputSchema, listJobLogsOutputSchema } from "@better-bull-board/core/jobs";
+import { listJobLogsInputSchema, listJobLogsOutputSchema } from "@better-bull-board/core/job-schemas";
 import { registerApiRoute } from "~/lib/utils/client";
 
 export const getJobLogsInput = listJobLogsInputSchema;
