+++ apps/app/src/app/api/jobs/table/schemas.ts
@@ -1,4 +1,4 @@
-import { listJobsInputSchema, listJobsOutputSchema } from "@better-bull-board/core/jobs";
+import { listJobsInputSchema, listJobsOutputSchema } from "@better-bull-board/core/job-schemas";
 import { registerApiRoute } from "~/lib/utils/client";
 
 export const getJobsTableInput = listJobsInputSchema;
