+++ packages/core/tsdown.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsdown";
 
 export default defineConfig({
-  entry: ["./src/jobs.ts", "./src/mutations.ts", "./src/overview.ts", "./src/queues.ts"],
+  entry: ["./src/job-schemas.ts", "./src/jobs.ts", "./src/mutations.ts", "./src/overview.ts", "./src/queues.ts"],
   format: ["esm"],
   dts: true,
   minify: true,
